完成代码模块化
