+++ src/SecondContent.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import 'antd/dist/reset.css';
+import "./App.css"
+import {  Divider } from 'antd';
+import Exam from './Exam.js'
+import Question from './Question.js';
+
+class SecondContent extends React.Component{
+    constructor(props){
+      super(props)
+    }
+    render(){
+      return(
+        <div className='secondContent'>
+          <Question/>
+          <Divider/>
+          <Exam/>
+        </div>
+      );
+    }
+  }
+export default SecondContent;
